requirements.txt
Flask==2.3.2
gunicorn==20.1.0
mysql-connector-python==8.0.33
geopy==2.3.0
numpy==2.0.0
pandas==2.2.2
scikit-learn==1.5.1